icmp6: use default interface as the route interface
As it stands now if we do dhcp and slaac we'll use the slaac interface for any
routes advertised on the local network as those will be the first interfaces we
find.  However if we have a default interface, ie the one set up from the pxe
dhcp packet, we would prefer to use this one.  So when getting RA's, check to
see if our default interface is a v6 interface and matches the card we got the
RA on, and if so use that as the route interface.  This allows us to get traffic
from the dhcp interface instead of through the slaac interface.

Signed-off-by: Josef Bacik <jbacik@fb.com>