bootp: don't add multiple interfaces for the same address
Since we can sometimes configure an interface from the dhcp packet that is still
in the pxe configuration any subsequent dhcp requests could add duplicate
interfaces for the same address.  Fix this by checking to see if we already have
an interface configured for the address we got from the server and return.  Also
make these functions void as nobody uses the return value.  Thanks,

Signed-off-by: Josef Bacik <jbacik@fb.com>