net: add interfaces when we open a card
Since we've started adding addresses to the multicast filter in efi we need to
make sure that any addresses that are added before we call ->open are added as
well.  This can happen when we configure interfaces from the dhcp packet left in
the pxe config.  Do this in a helper function and have the two people that call
->open use the helper function instead.  Thanks,

Signed-off-by: Josef Bacik <jbacik@fb.com>