bootp: use the ipv* version given by the dhcp server
We were setting DNS_OPTION_PREVER_IPV* based on which version of the dhcp
request we go back.  However this would mean that we would mistakenly fall back
on ipv4 in an ipv6 only system, or vice versa.  For the sake of simplicity lets
only use the ip version for the dns server we got in our dhcp response.  Thanks,

Signed-off-by: Josef Bacik <jbacik@fb.com>