push/pop errno in initrd read file path
If you try to load an initrd from http and it errors out we will free the initrd
context but continue on because net_tcp_socket_close() will reset the grub_errno
as will grub_initrd_close().  So we'll lose the errno and return GRUB_ERR_NONE
instead of the original error.  Add push/pulls to the appropriate places so we
don't lose our errno.  Thanks,

Signed-off-by: Josef Bacik <jbacik@fb.com>