efi term: open input console exlusively
We've had issues with a bunch of our EFI machines getting random console input
from God knows where.  This stops our provisioning because the timeout countdown
stops as it thinks we pressed a button, which we most certaintly did not.  So
open the input console in exclusive mode to make sure only the guy behind the
wheel gets to type.  With this patch we can no provision these weird machines.
Thanks,

Signed-off-by: Josef Bacik <jbacik@fb.com>